fix: handle signout properly, extrapolate nav component
