+++ src/components/ui/navigation.tsx
@@ -1,0 +1,170 @@
+"use client";
+
+import { CircleUserRound, LogOut, Settings, Users } from "lucide-react";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+
+import {
+  MobileNav,
+  MobileNavHeader,
+  MobileNavMenu,
+  MobileNavToggle,
+  Navbar,
+  NavbarButton,
+  NavbarLogo,
+  NavBody,
+  NavItems,
+} from "@/components/ui/nav";
+import { createClient } from "@/utils/supabase/client";
+import { Sidebar, SidebarBody, SidebarLink } from "~/ui/sidebar";
+
+interface NavigationProps {
+  loggedIn: boolean;
+}
+
+export default function Navigation({ loggedIn }: NavigationProps) {
+  const router = useRouter();
+  const [isMobileNavOpen, setIsMobileNavOpen] = useState(false);
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+
+  const navItems = [
+    {
+      name: "Features",
+      link: "#features",
+    },
+    {
+      name: "Pricing",
+      link: "#pricing",
+    },
+    {
+      name: "Contact",
+      link: "#contact",
+    },
+  ];
+
+  async function handleSignOut() {
+    try {
+      const supabase = createClient();
+      await supabase.auth.signOut();
+
+      router.refresh();
+    } catch (error) {
+      console.error("Error signing out:", error);
+    }
+  }
+
+  const sidebarItems = [
+    {
+      label: "Groups",
+      href: "/groups",
+      icon: (
+        <Users className="h-5 w-5 shrink-0 text-neutral-700 dark:text-neutral-200" />
+      ),
+    },
+    {
+      label: "Profile",
+      href: "/profile",
+      icon: (
+        <CircleUserRound className="h-5 w-5 shrink-0 text-neutral-700 dark:text-neutral-200" />
+      ),
+    },
+    {
+      label: "Settings",
+      href: "/settings",
+      icon: (
+        <Settings className="h-5 w-5 shrink-0 text-neutral-700 dark:text-neutral-200" />
+      ),
+    },
+    {
+      label: "Logout",
+      href: "#",
+      icon: (
+        <LogOut className="h-5 w-5 shrink-0 text-neutral-700 dark:text-neutral-200" />
+      ),
+      onClick: handleSignOut,
+    },
+  ];
+
+  if (loggedIn) {
+    return (
+      <Sidebar open={sidebarOpen} setOpen={setSidebarOpen}>
+        <SidebarBody className="justify-between gap-10">
+          <div className="flex flex-1 flex-col overflow-x-hidden overflow-y-auto">
+            <div className="flex flex-row items-center gap-2">
+              <span className="font-medium text-black dark:text-white">
+                Tavolo
+              </span>
+            </div>
+            <div className="mt-8 flex flex-col gap-2">
+              {sidebarItems.map((link, idx) => (
+                <SidebarLink key={idx} link={link} />
+              ))}
+            </div>
+          </div>
+          <div></div>
+        </SidebarBody>
+      </Sidebar>
+    );
+  }
+
+  return (
+    <Navbar className="mt-8">
+      <NavBody>
+        <NavbarLogo />
+        <NavItems items={navItems} />
+        <div className="flex items-center gap-4">
+          <NavbarButton href="/login" variant="secondary">
+            Login
+          </NavbarButton>
+          <NavbarButton href="/create" variant="primary">
+            Create a Group
+          </NavbarButton>
+        </div>
+      </NavBody>
+
+      <MobileNav>
+        <MobileNavHeader>
+          <NavbarLogo />
+          <MobileNavToggle
+            isOpen={isMobileNavOpen}
+            onClick={() => setIsMobileNavOpen(!isMobileNavOpen)}
+          />
+        </MobileNavHeader>
+
+        <MobileNavMenu
+          isOpen={isMobileNavOpen}
+          onClose={() => setIsMobileNavOpen(false)}
+        >
+          {navItems.map((item, idx) => (
+            <a
+              key={`mobile-link-${idx}`}
+              href={item.link}
+              onClick={() => setIsMobileNavOpen(false)}
+              className="relative text-neutral-600 dark:text-neutral-300"
+            >
+              <span className="block">{item.name}</span>
+            </a>
+          ))}
+          <div className="flex w-full flex-col gap-4">
+            <NavbarButton
+              onClick={() => setIsMobileNavOpen(false)}
+              href="/login"
+              variant="primary"
+              className="w-full"
+            >
+              Login
+            </NavbarButton>
+            <NavbarButton
+              onClick={() => setIsMobileNavOpen(false)}
+              href="/create"
+              variant="primary"
+              className="w-full"
+            >
+              Create a Group
+            </NavbarButton>
+          </div>
+        </MobileNavMenu>
+      </MobileNav>
+    </Navbar>
+  );
+}
