+++ src/components/ui/sidebar.tsx
@@ -1,14 +1,17 @@
 "use client";
-import { cn } from "@/lib/utils";
-import Link, { LinkProps } from "next/link";
-import React, { useState, createContext, useContext } from "react";
-import { AnimatePresence, motion } from "motion/react";
 import { IconMenu2, IconX } from "@tabler/icons-react";
+import { AnimatePresence, motion } from "motion/react";
+import type { LinkProps } from "next/link";
+import Link from "next/link";
+import React, { createContext, use, useState } from "react";
+
+import { cn } from "@/lib/utils";
 
 interface Links {
   label: string;
   href: string;
   icon: React.JSX.Element | React.ReactNode;
+  onClick?: () => void;
 }
 
 interface SidebarContextProps {
@@ -21,15 +24,15 @@ const SidebarContext = createContext<SidebarContextProps | undefined>(
   undefined,
 );
 
-export const useSidebar = () => {
-  const context = useContext(SidebarContext);
+export function useSidebar() {
+  const context = use(SidebarContext);
   if (!context) {
     throw new Error("useSidebar must be used within a SidebarProvider");
   }
   return context;
-};
+}
 
-export const SidebarProvider = ({
+export function SidebarProvider({
   children,
   open: openProp,
   setOpen: setOpenProp,
@@ -39,20 +42,21 @@ export const SidebarProvider = ({
   open?: boolean;
   setOpen?: React.Dispatch<React.SetStateAction<boolean>>;
   animate?: boolean;
-}) => {
+}) {
   const [openState, setOpenState] = useState(false);
 
   const open = openProp !== undefined ? openProp : openState;
   const setOpen = setOpenProp !== undefined ? setOpenProp : setOpenState;
 
   return (
-    <SidebarContext.Provider value={{ open, setOpen, animate: animate }}>
+    // eslint-disable-next-line react/no-unstable-context-value
+    <SidebarContext value={{ open, setOpen, animate }}>
       {children}
-    </SidebarContext.Provider>
+    </SidebarContext>
   );
-};
+}
 
-export const Sidebar = ({
+export function Sidebar({
   children,
   open,
   setOpen,
@@ -62,28 +66,28 @@ export const Sidebar = ({
   open?: boolean;
   setOpen?: React.Dispatch<React.SetStateAction<boolean>>;
   animate?: boolean;
-}) => {
+}) {
   return (
     <SidebarProvider open={open} setOpen={setOpen} animate={animate}>
       {children}
     </SidebarProvider>
   );
-};
+}
 
-export const SidebarBody = (props: React.ComponentProps<typeof motion.div>) => {
+export function SidebarBody(props: React.ComponentProps<typeof motion.div>) {
   return (
     <>
       <DesktopSidebar {...props} />
       <MobileSidebar {...(props as React.ComponentProps<"div">)} />
     </>
   );
-};
+}
 
-export const DesktopSidebar = ({
+export function DesktopSidebar({
   className,
   children,
   ...props
-}: React.ComponentProps<typeof motion.div>) => {
+}: React.ComponentProps<typeof motion.div>) {
   const { open, setOpen, animate } = useSidebar();
   return (
     <>
@@ -103,13 +107,13 @@ export const DesktopSidebar = ({
       </motion.div>
     </>
   );
-};
+}
 
-export const MobileSidebar = ({
+export function MobileSidebar({
   className,
   children,
   ...props
-}: React.ComponentProps<"div">) => {
+}: React.ComponentProps<"div">) {
   const { open, setOpen } = useSidebar();
   return (
     <>
@@ -153,9 +157,9 @@ export const MobileSidebar = ({
       </div>
     </>
   );
-};
+}
 
-export const SidebarLink = ({
+export function SidebarLink({
   link,
   className,
   ...props
@@ -163,7 +167,7 @@ export const SidebarLink = ({
   link: Links;
   className?: string;
   props?: LinkProps;
-}) => {
+}) {
   const { open, animate } = useSidebar();
   return (
     <Link
@@ -172,6 +176,12 @@ export const SidebarLink = ({
         "group/sidebar flex items-center justify-start gap-2 py-2 font-medium",
         className,
       )}
+      onClick={(e) => {
+        if (link.onClick) {
+          e.preventDefault();
+          link.onClick();
+        }
+      }}
       {...props}
     >
       {link.icon}
@@ -187,4 +197,4 @@ export const SidebarLink = ({
       </motion.span>
     </Link>
   );
-};
+}
